chore: move nextjs example wip
