--- examples/nextjs-api-reference/.eslintrc.cjs
@@ -1,3 +1,0 @@
-module.exports = {
-  extends: ['../../.eslintrc.cjs', 'plugin:@next/next/recommended'],
-}
