--- examples/nextjs-api-reference/app/client/components/Button.tsx
@@ -1,13 +1,0 @@
-'use client'
-
-import { useApiClientModal } from '@scalar/api-client-react'
-
-export const Button = () => {
-  const client = useApiClientModal()
-
-  return (
-    <button onClick={() => client?.open()}>
-      Click me to open the Api Client
-    </button>
-  )
-}
