--- examples/nextjs-api-reference/next.config.mjs
@@ -1,13 +1,0 @@
-/** @type {import('next').NextConfig} */
-const nextConfig = {
-  redirects: () => [
-    {
-      source: '/',
-      destination: '/reference',
-      permanent: true,
-    },
-  ],
-  output: 'standalone',
-}
-
-export default nextConfig
