--- examples/nextjs-api-reference/package.json
@@ -1,41 +1,0 @@
-{
-  "name": "@scalar-examples/nextjs-api-reference",
-  "license": "MIT",
-  "author": "Scalar (https://github.com/scalar)",
-  "homepage": "https://github.com/scalar/scalar",
-  "bugs": "https://github.com/scalar/scalar/issues/new/choose",
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/scalar/scalar.git",
-    "directory": "examples/nextjs-api-reference"
-  },
-  "version": "0.1.0",
-  "private": true,
-  "engines": {
-    "node": ">=18"
-  },
-  "scripts": {
-    "build": "next build",
-    "dev": "next dev -p 5058",
-    "docker:build": "build --platform=linux/amd64 -t ${image_name} --build-arg=\"BASE_IMAGE=scalar-base\" -f ./examples/nextjs-api-reference/Dockerfile .",
-    "lint:check": "next lint",
-    "lint:fix": "next lint --fix",
-    "start": "next start",
-    "types:check": "tsc --noEmit"
-  },
-  "type": "module",
-  "dependencies": {
-    "@scalar/api-client-react": "workspace:*",
-    "@scalar/api-reference-react": "workspace:*",
-    "@scalar/nextjs-api-reference": "workspace:*",
-    "next": "^15.1.1",
-    "react": "^19.0.0",
-    "react-dom": "^19.0.0"
-  },
-  "devDependencies": {
-    "@next/eslint-plugin-next": "^15.1.1",
-    "@types/node": "^20.17.10",
-    "@types/react": "^19.0.1",
-    "@types/react-dom": "^19.0.2"
-  }
-}
