+++ integrations/nextjs/playground/app/another-client/page.tsx
@@ -1,0 +1,14 @@
+import Link from 'next/link'
+
+import { Button, ClientWrapper } from '../client/components'
+
+const Page = () => {
+  return (
+    <ClientWrapper>
+      <Button />
+      <Link href="/client">Go to client</Link>
+    </ClientWrapper>
+  )
+}
+
+export default Page
