+++ integrations/nextjs/playground/next.config.mjs
@@ -1,0 +1,13 @@
+/** @type {import('next').NextConfig} */
+const nextConfig = {
+  redirects: () => [
+    {
+      source: '/',
+      destination: '/reference',
+      permanent: true,
+    },
+  ],
+  output: 'standalone',
+}
+
+export default nextConfig
